third_party/blink/web_tests/virtual/omt-worker-fetch/external/wpt/content-security-policy/inside-worker/shared-script-expected.txt

This is a testharness.js-based test.
PASS Cross-origin `importScripts()` not blocked in http:
PASS `eval()` not blocked in http:
PASS `setTimeout([string])` not blocked in http:
FAIL Cross-origin `importScripts()` blocked in http:?pipe=sub|header(Content-Security-Policy,script-src%20%27self%27 assert_throws: importScripts should throw `NetworkError` function "_ => importScripts("http://www.web-platform.test:8081/content-security-policy/support/var-a.js")" did not throw
FAIL `eval()` blocked in http:?pipe=sub|header(Content-Security-Policy,script-src%20%27self%27 assert_throws: `eval()` should throw 'EvalError'. function "_ => eval("1 + 1")" did not throw
FAIL `setTimeout([string])` blocked in http:?pipe=sub|header(Content-Security-Policy,script-src%20%27self%27 assert_equals: expected 0 but got 1
FAIL Cross-origin `importScripts()` blocked in http:?pipe=sub|header(Content-Security-Policy,default-src%20%27self%27 assert_throws: importScripts should throw `NetworkError` function "_ => importScripts("http://www.web-platform.test:8081/content-security-policy/support/var-a.js")" did not throw
FAIL `eval()` blocked in http:?pipe=sub|header(Content-Security-Policy,default-src%20%27self%27 assert_throws: `eval()` should throw 'EvalError'. function "_ => eval("1 + 1")" did not throw
FAIL `setTimeout([string])` blocked in http:?pipe=sub|header(Content-Security-Policy,default-src%20%27self%27 assert_equals: expected 0 but got 1
Harness: the test ran to completion.
